+++ paper/src/main/java/com/jannik_kuehn/loritimepaper/LoriTimePaper.java
@@ -111,7 +111,7 @@ public class LoriTimePaper extends JavaPlugin {
         loriTimePlugin.disable();
     }
 
-    public PaperPluginMessenger getBukkitPluginMessenger() {
+    public PaperPluginMessenger getPaperPluginMessenger() {
         return paperPluginMessenger;
     }