+++ CHANGELOG.md
@@ -6,6 +6,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] - ${maven.build.timestamp}
 ### Added
+- Paper plugin loader
+- Folia platform loader
 ### Changed
 ### Deprecated
 ### Removed